Mummi and bilbo contribution!! veriverse contribution!! LCTV contribution!! catherinacortega, euphemisminc, magrib, marianyc, purpleizlegitx, silvaesouza, sjpablo, thiagovgamg, wanket contribution!!

+++ appseed/aura/primitive/geometry/geometry_size.inl
@@ -341,14 +341,14 @@ inline size64::operator SIZE() throw()
 }
 
 inline size size::operator /(int i) { return size(cx / i, cy / i); }
-inline size size::operator /(double d) { return size(cx / d, cy / d); }
+inline size size::operator /(double d) { return size((int64_t)(cx / d), (int64_t)(cy / d)); }
 
 inline size size::operator *(int i) { return size(cx * i, cy * i); }
-inline size size::operator *(double d) { return size(cx * d, cy * d); }
+inline size size::operator *(double d) { return size((int64_t)(cx * d), (int64_t)(cy * d)); }
 
 inline size size::operator +(int i) { return size(cx + i, cy + i); }
-inline size size::operator +(double d) { return size(cx + d, cy + d); }
+inline size size::operator +(double d) { return size((int64_t)(cx + d), (int64_t)(cy + d)); }
 
 inline size size::operator -(int i) { return size(cx - i, cy - i); }
-inline size size::operator -(double d) { return size(cx - d, cy - d); }
+inline size size::operator -(double d) { return size((int64_t)(cx - d), (int64_t)(cy - d)); }
